requirements.txt
alembic==1.13.1
allure-pytest==2.13.3
allure-python-commons==2.13.3
anyio==4.6.2.post1
Appium-Python-Client==3.1.1
assertpy==1.1
attrs==23.1.0
browsermob-proxy==0.8.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.2.0
chromedriver-autoinstaller==0.6.4
click==8.1.7
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
databricks-cli==0.17.7
databricks-connect==13.2.1
databricks-sdk==0.6.0
databricks-sql-connector==3.1.0
dnspython==2.4.2
et-xmlfile==1.1.0
exceptiongroup==1.1.2
execnet==2.0.2
fonttools==4.48.1
google-auth==2.29.0
googleapis-common-protos==1.60.0
googletrans==4.0.0rc1
greenlet==2.0.2
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
hyperframe==5.2.0
idna==3.4
iniconfig==2.0.0
kiwisolver==1.4.5
logging==0.4.9.6
lz4==4.3.2
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.8.2
MouseInfo==0.1.3
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.1.4
pillow==10.2.0
playwright==1.44.0
pluggy==1.5.0
protobuf==4.24.1
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyarrow==14.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.21
pyee==11.1.0
PyGetWindow==0.0.9
PyJWT==2.8.0
pymongo==4.6.2
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyodbc==4.0.39
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pyspark==3.4.1
pytest==8.2.0
pytest-order==1.2.0
pytest-sqlalchemy==0.2.1
pytest-xdist==3.5.0
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
rubicon-objc==0.4.7
selenium==4.20.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
tabulate==0.9.0
thrift==0.16.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.16
urllib3-secure-extra==0.1.0
wsproto==1.2.0
yolk3k==0.9
